[#454] Split out mailer functionality
